Change LSS fastscan to work with multiple unknown devices
If there are multiple unknown devices and a second device sends a response (not sure if that is actually supposed to happen, but does seem to sometimes) then the second devices response to the 1st fastscan command gets incorrectly interpreted as the the 2nd fastscan command response. This causes the wrong bits to be set and the validate subsequently fails. Therefore always wait for the timeout before sending next LSS fastscan command